io-app-backend/requirements.txt
absl-py==2.3.1
aenum==3.1.16
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cligj==0.7.2
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
dask==2025.12.0
diffusers==0.36.0
docstring_parser==0.17.0
earthengine-api==1.7.4
einops==0.8.1
filelock==3.20.2
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
geopandas==1.1.2
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface_hub==1.2.4
hydra-core==1.3.2
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonargparse==4.45.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
lazy_loader==0.4
lightly==1.5.22
lightly-utils==0.0.2
lightning==2.6.0
lightning-utilities==0.15.2
locket==1.0.0
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
multidict==6.7.0
networkx==3.6.1
numpy==2.2.6
odc-geo==0.5.0
odc-loader==0.6.0
odc-stac==0.5.0
omegaconf==2.3.0
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
partd==1.4.2
pillow==12.1.0
planetary-computer==1.0.0
platformdirs==4.5.1
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.11
pydantic==2.12.5
pydantic_core==2.41.5
pyogrio==0.12.1
pyparsing==3.3.1
pyproj==3.7.2
pystac==1.14.2
pystac-client==0.9.0
python-box==7.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-lightning==2.6.0
pytz==2025.2
PyYAML==6.0.3
rasterio==1.5.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rioxarray==0.20.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
segmentation_models_pytorch==0.5.0
sentry-sdk==2.49.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.12
six==1.17.0
smmap==5.0.2
stringzilla==4.6.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
terratorch==1.2.1
threadpoolctl==3.6.0
tifffile==2025.12.20
timm==1.0.24
toolz==1.1.0
torch==2.9.1
torchgeo==0.8.0
torchmetrics==1.8.2
torchvision==0.24.1
tqdm==4.67.1
typer-slim==0.21.1
typeshed_client==2.8.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
wandb==0.23.1
Werkzeug==3.1.5
xarray==2025.12.0
yarl==1.22.0
zipp==3.23.0
